requirements.txt
streamlit
opencv-python
dlib
tensorflow
numpy
Pillow
scikit-learn
matplotlib
seaborn